financial_document_agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "financial-document-agent"
version = "0.1.0"
description = "Production-grade Python agent for analyzing banking documents"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Financial Agent Team" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pypdfium2>=4.30.0",
    "Pillow>=10.4.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
ocr-fallback = ["pytesseract>=0.3.13"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
]

[project.scripts]
financial-agent = "financial_agent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/financial_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src/financial_agent --cov-report=term-missing"
